Match the matrix's aspect ratio to the gameboard
Unfortunately, it turns out this was a bad idea. Depending on the
window's size, the gameboard can end up with way, *way* too many cells,
and it crushes the browser under its heel.

The bounds-checking in setAspectRatio proved crucial; when one or both
dimensions is zero the aspect ratio is Infinity and it tries to make an
infinite matrix. Problematic.