Update header elements for narrow screens
The `min-width: max-content` ensures the app background stretches to
contain the gameboard when it's wider than the viewport.

Undid some of the squishing in 9a6684f5fc8526b5c26556349c930fc5ab2998f1,
since now that the viewport is scrollable you can just move it around if
you need to see more controls.